extract todo input component
